z 128 na 64

+++ moneybag.h
@@ -70,8 +70,8 @@ public:
     // obiektu Value z liczbą całkowitą nieujemną
     std::strong_ordering operator<=>(Value const &) const = default;
     constexpr bool operator==(Value const &) const = default;
-    std::strong_ordering operator<=>(const __uint128_t &) const;
-    constexpr bool operator==(const __uint128_t &) const;
+    std::strong_ordering operator<=>(const uint64_t &) const;
+    constexpr bool operator==(const uint64_t &) const;
     // Wypisywanie wartości
     explicit operator std::string () const;
